# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Oops AI-native Software Development Workspace API",
        "title": "Oops API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
        "/auth/register": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Register a user account",
                "parameters": [
                    {
                        "description": "user data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_modules_identity_interface.createUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_identity_interface.userResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_oopsla5xx_oops-api-v1_internal_shared_response.Response"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/github_com_oopsla5xx_oops-api-v1_internal_shared_response.Response"
                        }
                    }
                }
            }
        },
        "/health": {
            "get": {
                "description": "Returns service liveness status. Does NOT check DB/Redis — use this for liveness probes only.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Health check",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_modules_health_interface.healthResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "github_com_oopsla5xx_oops-api-v1_internal_shared_response.ErrorInfo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "field": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "github_com_oopsla5xx_oops-api-v1_internal_shared_response.Meta": {
            "type": "object",
            "properties": {
                "page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "github_com_oopsla5xx_oops-api-v1_internal_shared_response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/github_com_oopsla5xx_oops-api-v1_internal_shared_response.ErrorInfo"
                },
                "meta": {
                    "$ref": "#/definitions/github_com_oopsla5xx_oops-api-v1_internal_shared_response.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "internal_modules_health_interface.healthResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "internal_modules_identity_interface.createUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string",
                    "maxLength": 72
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "internal_modules_identity_interface.userResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "first_name": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_name": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}